test: InputView와 RacingCarService로 테스트 분리
