+++ src/test/java/racingcar/ApplicationTest.java
@@ -33,80 +33,6 @@ class ApplicationTest extends NsTest {
         );
     }
 
-    @Test
-    void 자동차_이름이_공백일_때_예외_테스트() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException(" ", "1"))
-                        .isInstanceOf(IllegalArgumentException.class)
-                        .hasMessageContaining("자동차 이름은 공백이 될 수 없습니다.")
-        );
-    }
-
-    @Test
-    void 자동차_이름_5자_초과_예외_테스트() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("jijijiji, coco", "1"))
-                        .isInstanceOf(IllegalArgumentException.class)
-                        .hasMessageContaining("자동차 이름은 5자 이하로 입력해 주세요.")
-        );
-    }
-
-    @Test
-    void 자동차_이름_중복_입력_예외_테스트() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("jiji, jiji", "1"))
-                        .isInstanceOf(IllegalArgumentException.class)
-                        .hasMessageContaining("자동차 이름은 중복될 수 없습니다.")
-        );
-    }
-
-    @Test
-    void 횟수_음수_입력_예외_테스트() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi,woni", "-1"))
-                        .isInstanceOf(IllegalArgumentException.class)
-                        .hasMessageContaining("게임 횟수는 양수를 입력해 주세요.")
-        );
-    }
-
-    @Test
-    void 횟수_문자열_입력_예외_테스트() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi,woni", "one"))
-                        .isInstanceOf(IllegalArgumentException.class)
-                        .hasMessageContaining("게임 횟수는 양수를 입력해 주세요.")
-        );
-    }
-
-    @Test
-    void 횟수_0_입력_예외_테스트() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi,woni", "0"))
-                        .isInstanceOf(IllegalArgumentException.class)
-                        .hasMessageContaining("게임 횟수는 양수를 입력해 주세요.")
-        );
-    }
-
-    @Test
-    void 횟수_공백_입력_예외_테스트() {
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> runException("pobi,woni", " "))
-                        .isInstanceOf(IllegalArgumentException.class)
-                        .hasMessageContaining("게임 횟수는 공백이 될 수 없습니다.")
-        );
-    }
-
-    @Test
-    void 문자열_분리_테스트() {
-        String carNameList = "jiji, kiki , didi";
-
-        RacingCarService racingCarService = new RacingCarService();
-        Set<String> carNames = racingCarService.splitCarName(carNameList);
-
-        assertThat(carNames).hasSize(3);
-        assertThat(carNames).containsExactlyInAnyOrder("jiji", "kiki", "didi");
-    }
-
     @Override
     public void runMain() {
         Application.main(new String[]{});
